+++ core/data/src/main/kotlin/fr/forumhfr/redface2/core/data/di/PlatformBindingsModule.kt
@@ -8,6 +8,7 @@ import fr.forumhfr.redface2.core.domain.coroutines.DefaultDispatcher
 import fr.forumhfr.redface2.core.domain.coroutines.IoDispatcher
 import fr.forumhfr.redface2.core.domain.coroutines.MainDispatcher
 import fr.forumhfr.redface2.core.parser.HfrParser
+import fr.forumhfr.redface2.core.parser.flags.FlagsListParser
 import fr.forumhfr.redface2.core.parser.messages.PrivateMessageListParser
 import java.time.Clock
 import javax.inject.Singleton
@@ -43,4 +44,8 @@ object PlatformBindingsModule {
     @Provides
     @Singleton
     fun providePrivateMessageListParser(): PrivateMessageListParser = PrivateMessageListParser()
+
+    @Provides
+    @Singleton
+    fun provideFlagsListParser(): FlagsListParser = FlagsListParser()
 }